refactor: move store to stores folder
